+++ server/cart/src/main/kotlin/com/kenlikdev/qmarket/cart/repository/CartRepository.kt
@@ -2,9 +2,8 @@ package com.kenlikdev.qmarket.cart.repository
 
 import com.kenlikdev.qmarket.cart.domain.Cart
 import org.springframework.data.jpa.repository.JpaRepository
-import java.util.Optional
 import java.util.UUID
 
 interface CartRepository : JpaRepository<Cart, UUID> {
-    fun findByUserId(userId: UUID): Optional<Cart>
+    fun findByUserId(userId: UUID): Cart?
 }